+++ algonaut_indexer/src/models/lookup_account_transactions_200_response.rs
@@ -12,7 +12,7 @@
 pub struct LookupAccountTransactions200Response {
     /// Round at which the results were computed.
     #[serde(rename = "current-round")]
-    pub current_round: i32,
+    pub current_round: u64,
     /// Used for pagination, when making another request provide this token with the next parameter.
     #[serde(rename = "next-token", skip_serializing_if = "Option::is_none")]
     pub next_token: Option<String>,
@@ -22,7 +22,7 @@ pub struct LookupAccountTransactions200Response {
 
 impl LookupAccountTransactions200Response {
     pub fn new(
-        current_round: i32,
+        current_round: u64,
         transactions: Vec<crate::models::Transaction>,
     ) -> LookupAccountTransactions200Response {
         LookupAccountTransactions200Response {
